requirements.txt
Django==1.11
django-auth-ldap==1.2.8
django-extensions==2.0.0
psycopg2==2.7.4
six==1.10.0
virtualenv==15.0.2
python-ldap==2.4.25
django-simple-history==1.9
pytz==2017.2
docutils==0.14
python-decouple==3.1
django-admin-view-permission==1.6
factory-boy==2.11.1
django-rest-swagger==2.2.0
django-ajax-selects==1.7.1
django-model-utils==3.1.2
django-modeladmin-reorder==0.3.1
django-grappelli==2.10.4
django-nested-admin==3.1.2
parameterized==0.6.1

django-cors-headers==2.4.0

# API
djangorestframework==3.7.7
markdown==2.6.11
django-filter==1.1.0
coreapi==2.3.3


coverage==4.4.1
pytest-django==3.1.2
pytest-cov==2.5.1
codacy-coverage==1.3.6
pylint==1.7.4
pylint-django==0.7.2
# Additional libs for python-ldap: http://stackoverflow.com/questions/4768446/i-cant-install-python-ldap

# Docker
#docker-compose==1.7.1
#docker-py==1.8.1
#dockerpty==0.4.1
tblib==1.3.2